Add files via upload
The holiday balance was altered by customers using the GUI we provided and the current Database would lose track of the total. I created this to use correct the data while we waited for the scripts to be written again. Issues ranged from booking 2 holidays on the same day, starting the holiday year 05/01/20xx to 31/12/20xx so holiday would not count between the 1-4 of January, I then proceeded to add in readable error messages as to why the data was wrong and inserted it into a temp table should I want t change the original table